+++ NEWS.md
@@ -4,6 +4,9 @@
 
 * Function for generating synthetic data (generate_synthetic_data)
 * Unittest for generate_synthetic_data
+* Function to compute residuals and unittest
+* Function to impute NA values based on density and unittest
+* Function to separate prediction model training
 
 ## Changed